pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "django-ninja-fsbr"
authors = [{ name = "Fabian Binz", email = "fabian.binz@gmail.com" }]
dynamic = ["version", "description"]
dependencies = ["django-ninja"]

[project.urls]
Home = "https://github.com/fbinz/django-ninja-fsbr"

[tool.flit.module]
name = "ninja_fsbr"
